test(sbs3): share one Spring context across details integration tests
Both details integration tests now extend a single AbstractDetailsSharedContextTest
base that owns the unioned context configuration (web environment, properties,
imports and shared beans), so the whole package resolves one cached
ApplicationContext instead of each test building its own.
